% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ACNDiversityPlot}
\alias{ACNDiversityPlot}
\title{Absolute Copy-Number Diversity Heatmap}
\usage{
ACNDiversityPlot(
  long_segments = data.frame(),
  order = TRUE,
  ret_order = FALSE,
  save_path = FALSE,
  rmblacklist = NULL,
  addtitle = NULL,
  ann_df = NULL,
  obj_name = "version1"
)
}
\arguments{
\item{long_segments}{A dataframe. This should be in long format with the same number of rows (bins) for each sample. \cr
If using the output from the \code{CalculateACNs} function then first run \code{SegmentsToCopyNumber} to convert to long. \cr
Expects the following columns: \cr
c("chromosome", "start", "end", "state", "sample_id")}

\item{order}{(optional) Character vector. Defines the order in which samples will be plotted. \cr
This is particularly useful when plotting more than 1 heatmap next to one another. \cr
Allows samples to line-up horizontally.
Example: \cr
\verb{c("CC-CHM-1341", "CC-CHM-1347", "CC-CHM-1355", CC-CHM-1361", "CC-HAM-0369", "CC-HAM-0374", "CC-HAM-0379", "CC-HAM-0383", "CC-HAM-0385")}}

\item{ret_order}{(optional) Logical. If set to TRUE the function returns the order in which samples were plotted.}

\item{save_path}{(optional) String. Expects a path to a directory where the plot should be saved. (png)}

\item{rmblacklist}{optional \emph{NULL or character vector}. Filter the plot of blacklist regions from this genome. ex. 'hg19'}

\item{obj_name}{(optional) String. Adds a tag to the end of the filename if saving the image. \cr
Only used if the save_path parameter is also set.}
}
\value{
A ggplot2 object or a list of a ggplot2 object and vector of the ordered sample names.
}
\description{
A whole genome heatmap of copy-number changes. It expects absolute copy-numbers.
Default behaviour is to perform hierarchical clustering of the samples and plot the heatmap in that order.
Setting the \code{order} parameter overrules this behaviour.
}
