% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ACNSegmentsPlot}
\alias{ACNSegmentsPlot}
\title{Plot Absolute Copy-Number Segments}
\usage{
ACNSegmentsPlot(x, sample, sample_segments, sample_cns, output)
}
\description{
Plot Absolute Copy-Number Segments
}
