% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_gather_segments_ratiomedian.R
\name{AddChromosomeArm}
\alias{AddChromosomeArm}
\title{Adds a column for the chromosome arm in which each segment is in.}
\usage{
AddChromosomeArm(bam_ratios_frame, include_chr_X)
}
\arguments{
\item{bam_ratios_frame}{A data frame: the cleaned up version of the bam_ratios.txt file. In the algorithm,
this segment data has already had its spurious regions removed.}

\item{include_chr_X}{A boolean: True if we include chromosome X/23, False otherwise.}
}
\description{
More specifically, iterates through all the segments, and using genomic data for the
middle positions of all chromosomes, it determines whether the segment is to the left
or to the right of the middle position. For example, if a segment is in chromosome 5
and is to the left of the middle position, it is in chromosome arm 1. If it were to the right
instead, then its chromosome arm would be 1.5.

Note that most of this iterative process is done in the AddChromosomeArmHelper helper.
}
