% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{AddGenesToPlot}
\alias{AddGenesToPlot}
\title{Label genes of interest on an absolute copy-number plot.}
\usage{
AddGenesToPlot(plot, genes, edb = EnsDb.Hsapiens.v75, ...)
}
\arguments{
\item{plot}{A ggplot object, corresponding to the existing absolute copy-number plot.}

\item{genes}{A vector of gene symbols you'd like to label on the plot -- use upper case.}

\item{edb}{The name of the local Ensembl DB package to use for looking up gene information. Defaults to \link{EnsDb.Hsapiens.v75}, which annotates hg19/GRCh37.}

\item{...}{Additional options controlling the visual display of the labels; passed to ggrepel::geom_label_repel(), and supports any of its arguments.}
}
\value{
A ggplot object with the gene labels added.
}
\description{
This function takes an existing absolute copy-number plot, a vector of gene names, and optionally an Ensembl DB to use for looking up gene annotations.
Ensure that you use an Ensembl DB which annotates the same genome version used for your existing plot.
A plot with the genes of interest labeled is returned.
}
\seealso{
\code{\link[ggrepel:geom_text_repel]{ggrepel::geom_label_repel()}} for supported arguments that alter the visual display of the labels.
}
