% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_levels.R
\name{AssignLevels}
\alias{AssignLevels}
\title{Gather big segments and assign levels iteratively}
\usage{
AssignLevels(segments, thr)
}
\arguments{
\item{segments}{A data frame containing segment data.}

\item{thr}{A float: the threshold previously estimated via KDE. Used to determine whether
large segments are merged.}
}
\description{
Iterates through the given segments_copy data frame. Does the following at each iteration:
\enumerate{
\item Finds the largest segment (by size). Then, finds all the segments 'close' to it: closeness
here means that that ratio_median difference between them is less than threshold (\code{thr}).
\item Once we identify these segments that are 'close' to each other (called closest_indices),
we iterate through them, and set their level in \code{segments}.
\item Finally, remove these closest_indices segments from \code{segments_copy}.
}
}
