% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_call_lga.R
\name{BreakSmoothToLGA}
\alias{BreakSmoothToLGA}
\title{Preps \code{segments} for LGA call.}
\usage{
BreakSmoothToLGA(threshold, segments, granges_obj)
}
\arguments{
\item{threshold}{A float. Threshold in ratio_median difference previously estimated.}

\item{segments}{A data frame: segment data that's (ideally) already been processed (merging segments,
small ones have been reinserted, and so on)}

\item{granges_obj}{A GRanges object.}
}
\description{
Does many things, details found within function body. First, assigns segments
that meet certain criteria an index of 0 (there's 4 cases + edge cases for assigning an
index of 0, inside the inner for-loop.) Then, removes the segments that were assigned this
index of 0. They're considered "small breaks".
}
