% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wisecondorXtoQDNAseq.R
\name{BuildWxQdnaObject}
\alias{BuildWxQdnaObject}
\title{Build a QDNAseq object from the WisecondorX output files}
\usage{
BuildWxQdnaObject(
  input_path,
  genome_used = "hg19",
  bin_size = 30000,
  Xincluded = FALSE
)
}
\arguments{
\item{input_path}{(string) The path to the output files generated by WisecondorX. A directory.}

\item{genome_used}{(string) The reference genome used for alignment. (default: hg19)}

\item{bin_size}{(integer) The bin size used during copy number calling in base pairs. (default: 30000)}

\item{Xincluded}{(logical) Whether or not the X chromosome was included in this analysis. (default: FALSE)}
}
\value{
A QDNAseq object.
}
\description{
In order to do 1-to-1 comparisons between QDNAseq and WisecondorX outputs it is useful to have outputs in the same format.
This function creates a QDNAseq object out of the WisecondorX output files and returns it.
}
