% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_call_lga.R
\name{CallLGA}
\alias{CallLGA}
\title{Determines number of large genomic alterations (LGAs) for sizes 3:11}
\usage{
CallLGA(threshold, segments)
}
\arguments{
\item{threshold}{A float. Threshold in ratio_median difference previously estimated.}

\item{segments}{A data frame: segment data that's (ideally) already been processed (merging segments,
small ones have been reinserted, and so on)}
}
\description{
Uses DetermineNumberOfLGAs to determine the number of LGAs for sizes 3 to 11 (Mb).
}
