% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signaturegeneration.R
\name{CallSignatureExposures}
\alias{CallSignatureExposures}
\title{Wrapper to find Signature Exposures for a set of samples}
\usage{
CallSignatureExposures(
  copy_numbers_input,
  component_models = NULL,
  signatures = NULL,
  data_path = NULL,
  plot_savepath = NULL,
  sigs_savepath = NULL,
  relativeCN_data = NULL,
  refgenome = "hg19"
)
}
\arguments{
\item{copy_numbers_input}{A list of dataframes. Each dataframe should be the segmented copy-numbers of a sample.}

\item{component_models}{String. The mixture models to use in determining the components present in this dataset. \cr
Options: \cr
'component_models/30kb_ovarian/component_models_britroc_aCNs.rds' - Components determined by modelling the Britroc Ovarian sWGS samples. (Macintyre 2018) \cr
'vancouver_HQendo_VAFrascal' - Components determined by modelling high-quality p53abn sWGS endometrial samples. (2022) \cr}

\item{signatures}{String. The signatures themselves. Please name those for which to calculate exposures. \cr
Options: \cr
'signatures/30kb_ovarian/component_by_signature_britroc_aCNs.rds' - The signatures created using the Britroc Ovarian sWGS samples. (Macintyre 2018) \cr
'VAFrascalHQendoVan6sigs' - The signatures created using the high-quality p53abn sWGS endometrial samples. \cr}

\item{data_path}{String. The path where to find signatures objects and component models.}

\item{plot_savepath}{(optional) String. The path where to save a sample-by-component heatmap. Please provide a directory.}

\item{sigs_savepath}{(optional) String. The path where to save a csv of the calculated signature exposures. Please provide a directory.}

\item{relativeCN_data}{(optional) Logical. If using relative Copy-Number data as input, set to true. Otherwise, ignore.}

\item{refgenome}{(optional) String. The reference genome used. (ex. 'hg19', or 'hg38')}
}
\value{
A dataframe of the exposures for each sample to each signature.
}
\description{
CallSignatures() calculates copy-number signature exposures for a list of samples given their segmented copy-number profiles.
This function makes use of code found in the CN-Signatures bitbucket repo.
and described more robustly here https://www.nature.com/articles/s41588-018-0179-8.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{todo - make examples
}\if{html}{\out{</div>}}
}
