% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ChromosomeSplit}
\alias{ChromosomeSplit}
\title{Relative copy number plotting functions}
\usage{
ChromosomeSplit(x)
}
\arguments{
\item{x}{\emph{dataframe} containing the copy number segments in long format: chr column with format chromosome:start-stop, sample and segVal}
}
\value{
dataframe containing three separate columns for the chromosome number, sample and segVal
}
\description{
Function to split chromosome names into the separate columns with the chromosome number, start and ending position
}
\details{
A QDNAseq object generally contains the chromosome and position information in the following format chromosome:start-end. This often needs to be separated into three columns for plotting relative copy numbers and other data wrangling.
}
