% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ChromosomeSplitPos}
\alias{ChromosomeSplitPos}
\title{Split Chromosome Name}
\usage{
ChromosomeSplitPos(x)
}
\arguments{
\item{x}{\emph{dataframe} containing the copy number segments in long format with columns: \cr
chr sample segVal}
}
\value{
dataframe containing three separate columns for the chromosome number, sample and segVal
}
\description{
Function to split compounded chromosome-position names into separate columns.
Distinct columns for chromosome number, start, and ending positions are the result.
}
\details{
A QDNAseq object generally contains the chromosome and position information as such: \cr
chromosome:start-end.
This is inconvenient for plotting and other data wrangling.
}
