% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_gather_segments_ratiomedian.R
\name{CleanBamRatiosFrame}
\alias{CleanBamRatiosFrame}
\title{Performs some cleaning operations.}
\usage{
CleanBamRatiosFrame(raw_bam_ratios, log_transform)
}
\arguments{
\item{raw_bam_ratios}{A data frame: the raw bam ratios file obtained from the bam_ratio.txt file.}

\item{log_transform}{A boolean: TRUE if you want to log2 transform the ratio_median column. FALSE otherwise.
If dealing with Controlfreec data, log_transform should be TRUE. If QDNAseq, log_transform should be FALSE.}
}
\description{
These pre-processing steps are:
\enumerate{
\item Replaces chromosome 'X' for '23'.
\item Removes segments in chromosome 'Y'.
\item Adds an extra column for the End Position of the segment.
\item Removes segments whose ratio and ratio_median are -1.
\item Log2 transforms the ratio and ratio_median columns.
\item Removes segments with a ratio_median = -Inf.
}
}
