% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{CNSegmentsPlot}
\alias{CNSegmentsPlot}
\title{Segment Copy Number Plot across the Genome}
\usage{
CNSegmentsPlot(
  cnobj,
  sample = NULL,
  copy_number_steps = NULL,
  max_points_to_display = Inf,
  highlight_masks = NULL,
  min_copy_number = NULL,
  max_copy_number = NULL,
  copy_number_breaks = NULL,
  dolog2 = NULL,
  def_point_colour = "black",
  point_alpha = 0.6,
  point_size = 1,
  point_shape = 16,
  segment_colour = "red",
  segment_alpha = 1,
  segment_line_size = 1.5,
  copy_number_step_colour = "blue",
  copy_number_step_alpha = 0.35,
  copy_number_step_line_size = 0.75,
  xlabel = "chromosome",
  ylabel = "copy number"
)
}
\arguments{
\item{cnobj}{A \emph{QDNAseq} object contains \code{copynumber}, and \code{segmented} assay slots. \cr
This function expects copy-number data that has been segmented. May contain multiple samples.}

\item{sample}{A \emph{character string} indicating the sample. \cr
Required if the object passed to \code{cnobj} contains data for multiple samples. \cr
Ex. \code{c('sample1')} \cr
Must match one of the sample names exactly.}

\item{copy_number_steps}{a data frame containing \code{absolute_copy_number}
and \code{copy_number} columns.}

\item{max_points_to_display}{maximum number of copy number points to display
(downsampling may be carried out if there are more copy number values than
this number).}

\item{highlight_masks}{A \emph{character string} indicating any feature data slots
containing logical masks over the bins. Each slot will be given its own colour. \cr
Keep in mind that NA values will be filtered out and should masks have overlap
the last one will take precedence. \cr
Ex. \code{c('comCNV.mask', 'centro.telo.mask')}}

\item{min_copy_number}{Numeric. The minimum \code{copy_number} to display. Effectively controls Y-axis limits.}

\item{max_copy_number}{Numeric. The maximum \code{copy_number} to display. Effectively controls Y-axis limits.}

\item{copy_number_breaks}{Vector of integers. Breaks at which grid lines will be displayed. \cr
ex. c(1:12)}

\item{def_point_colour}{A \emph{character string} indicating the colour for the
default copy number points.}

\item{point_alpha}{Numeric. The transparency of the copy number points.}

\item{point_size}{The size of the copy number points.}

\item{point_shape}{The point shape. See the ggplot2::geom_point() docs and
the \code{shape} parameter for details.}

\item{segment_colour}{The colour of the copy number segments.}

\item{segment_alpha}{The transparency of the copy number segments.}

\item{segment_line_size}{The size of the lines for copy number segments.}

\item{copy_number_step_colour}{The colour of the copy number step lines.}

\item{copy_number_step_alpha}{The transparency of the copy number step lines.}

\item{copy_number_step_line_size}{The size of the lines for the copy number.
steps.}

\item{xlabel, ylabel}{x- and y-axis labels.}
}
\value{
A \code{ggplot} object.
}
\description{
Whole genome copy number plot for one sample with partitioning lines between chromosomes.
}
\details{
The \code{copynumber} values passed into this function can be on the relative
or absolute scales, or be log2 ratios. However, the same scale should be used
consistently in each slot of \code{cnobj} (and the \code{copy_number_steps} dataframe as the case may be).
}
