% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplequality.R
\name{CollapsedSegs}
\alias{CollapsedSegs}
\title{Collapse segments}
\usage{
CollapsedSegs(x)
}
\arguments{
\item{x}{Dataframe containing the following columns: chromosome, start, stop, segmented, and copy_number}
}
\description{
Collapsed segmented dataframe
}
