% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFilteredCNSummaryTable.R
\name{CollapseTableToSegments}
\alias{CollapseTableToSegments}
\title{Collapses down table to segmented copy numbers}
\usage{
CollapseTableToSegments(df)
}
\arguments{
\item{df}{A dataframe. A dataframe with copy number calls (with columns 'chromosome', 'start', 'end', 'mean.cn')}
}
\value{
A dataframe. A dataframe of segment summaries of various characteristics (derived from copy-number calls)
}
\description{
CollapseTableToSegments transforms relative copy-number calls to segment tables.
}
