% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CopyNumberSegments}
\alias{CopyNumberSegments}
\title{Collapses relative copy-number calls to segment tables}
\usage{
CopyNumberSegments(copy_number)
}
\arguments{
\item{copy_number}{A dataframe. A dataframe with copy number calls (with columns 'sample_id', 'chromosome', 'start', 'end', 'segmented')}
}
\value{
A dataframe. A dataframe of summaries of various characteristics (derived from copy-number calls)
}
\description{
CopyNumberSegments() transforms relative copy-number calls to segment tables.
Inverse of the SegmentsToCopyNumber function.
}
