% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplequality.R
\name{CopySegFlat}
\alias{CopySegFlat}
\title{Make a DF of CNs and Segments from S4 QDNAseq object}
\usage{
CopySegFlat(x)
}
\description{
Function to create a dataframe containing both the relative copy numbers and the segmented calls from a QDNAseq object.
}
