% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_helpers_small_segments.R
\name{CorrectLeftovers}
\alias{CorrectLeftovers}
\title{Adds 'leftsovers' to the segment frame after re-inserting small segments.}
\usage{
CorrectLeftovers(segments, granges_obj)
}
\arguments{
\item{segments}{A data frame. Segments with size >= 3Mb.}

\item{granges_obj}{A GRanges object: is used as reference to check whenever we have
empty space between segments and get the ratio_median of this missing portion.}
}
\description{
Iterates over the \code{segments} frame and if the space between adjacent segments is
greater than one, then we use GRanges to obtain genomic data from this missing region in the
data frame. We update the adjacent segment's start or end positions depending on
which segment is closer to the missing segment.

Do this as long as adjacent segments are in the same chromosome arm.
}
