% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_helpers_call_lga.R
\name{DetermineNumberOfLGAs}
\alias{DetermineNumberOfLGAs}
\title{Determines if the given segments are LGAs of the given size}
\usage{
DetermineNumberOfLGAs(threshold, size_lga, segments)
}
\arguments{
\item{threshold}{A float. Threshold in ratio_median difference previously estimated.}

\item{size_lga}{An integer: size of the LGA to look for.}

\item{segments}{A data frame: segment data.}
}
\description{
Iterates through the segments in \code{segments} and firstly checks if the segment meets the definition
of an LGA:
\enumerate{
\item Segment and Previous are in the same chromosome arm.
\item Size of space between Segment and Previous is less than 3Mb.
Then, we check that:
\item Size of Segment >= \code{size_lga}
\item Size of Previous >= \code{size_lga}
Finally, we check that ratio_median difference between Segment and Previous > THR. If so, then mark segment as an LGA (via 1's in the output frame).
}
}
