% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_levels.R
\name{ExcludeShortArms}
\alias{ExcludeShortArms}
\title{Removes the short chromosome arms (21, 22, 24) & keeps segments with size >= 3Mb}
\usage{
ExcludeShortArms(segments)
}
\arguments{
\item{segments}{The segments data frame, gathered by ratio medians.}
}
\description{
Removes any segments that lie on a chromosome arm that:
\itemize{
\item is greater than 23.6 (i.e chromosome arms 24 and beyond)
\item is equal to 21 (i.e. first chromosome arm of chromosome 21)
\item is equal to 22 (i.e. first chromosome arm of chromosome 22)
Then, updates sizes of the kept segments.
}
}
