% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ExportBinsQDNAObj}
\alias{ExportBinsQDNAObj}
\title{Export the CN bins of a QDNAseq Object to a wide dataframe}
\usage{
ExportBinsQDNAObj(
  object,
  type = c("copynumber", "segments", "calls"),
  filter = TRUE,
  digits = 3,
  chromosomeReplacements = c(`23` = "X", `24` = "Y", `25` = "MT"),
  ...
)
}
\arguments{
\item{object}{\emph{QDNAseq Object} containing the copy number segments in long format with columns: \cr
chr sample_id segVal}

\item{type}{\emph{character} Type of data to export, options are "copynumber" (corrected or uncorrected read counts), "segments", or "calls".}

\item{filter}{\emph{logical} If @TRUE, bins are filtered, otherwise not.}

\item{digits}{\emph{numeric} The number of digits to round to.}

\item{chromosomeReplacements}{\emph{named char vector} A named character vector of chromosome name replacements to be done.
Only used when \code{object} is of class: \cr
\code{cghRaw}, \code{cghSeg}, \code{cghCall}, or \code{cghRegions}, \cr
since these classes store chromosome names as integers, whereas all QDNAseq object types use character vectors.
Defaults to \code{c("23"="X", "24"="Y", "25"="MT")} for human.}
}
\value{
A wide dataframe containing a column for each sample_id as well as 4 additional columns: \cr
\verb{feature, chromosome, start, end}
}
\description{
Function to export the bins of a QDNAseq object to a wide dataframe.
Can be any of the "copynumber", "segments", or "calls" slots.
Unlike the similar function 'exportBins' in the QDNAseq package it does not log-normalize.
}
