% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplequality.R
\name{FilterCNs}
\alias{FilterCNs}
\title{Remove or set a mask for relative CN bins (QDNAseq object)}
\usage{
FilterCNs(
  cnobj,
  filter_by = NULL,
  genome = "hg19",
  minimum_overlap = 5000,
  removebins = FALSE,
  maskgaps = FALSE,
  maskcomCNVs = FALSE,
  maskname = "mask1"
)
}
\arguments{
\item{cnobj}{\emph{QDNASeq object} containing the relative copy number calls as well as the segmented relative copy number calls}

\item{filter_by}{Optional \emph{Dataframe} containing genomic regions to filter/mask. One region expected per row. \cr
Expected columns: \cr
chr start end etc. etc.}

\item{genome}{A character \emph{string} of the reference genome used. ex "hg38"}

\item{minimum_overlap}{An \emph{integer} designating the number minimum number of overlapping bases for a bin to be removed.}

\item{removebins}{A \emph{Logical}. If TRUE then remove bins overlapping the indicated regions.}

\item{maskgaps}{A \emph{Logical}. If TRUE then remove bins overlapping the centromeric and telomeric regions. Be sure to indicate the correct reference genome.}

\item{maskcomCNVs}{A \emph{Logical}. If TRUE then remove bins overlapping commonly copy-number variant regions in this organism. Be sure to indicate the correct reference genome.}

\item{maskname}{A character \emph{string} for the name to be used in case of setting a mask. Set this field if providing a DF to the filter_by parameter.}
}
\value{
A QDNAseq object filtered or with masks set.
}
\description{
Remove or set a mask for relative CN bins
}
\details{
Expects the input of a QDNAseq object containing relative copy-number calls (such as from QDNASeq or WiseCondorX). \cr
}
