% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpersignaturefunctions.R
\name{FitComponent}
\alias{FitComponent}
\title{Fit a Mixture Model for a provided CN-Feature}
\usage{
FitComponent(
  dat,
  dist = "norm",
  seed = 77777,
  model_selection = "BIC",
  min_prior = 0.001,
  niter = 1000,
  nrep = 1,
  min_comp = 2,
  max_comp = 10
)
}
\value{
A flexmix object.
}
\description{
Fit a Mixture Model of gaussians or poissons for a provided CN-Feature.
}
