% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicimportance.R
\name{FitGIModelsCNtoS}
\alias{FitGIModelsCNtoS}
\title{Fit a model to each bin in a CN-dataset predicting signature exposure}
\usage{
FitGIModelsCNtoS(
  X,
  Y,
  model_option = "beta",
  bin_indices = NULL,
  data_slot = NULL,
  run_parallel = FALSE,
  cores = 1,
  epsilon = 5e-04
)
}
\arguments{
\item{X}{\emph{S4} A QDNAseq object that contains \code{copynumber}, and/or \code{segmented} assay slots. \cr
OR alternatively a matrix of bins (x) by samples (y) may be provided. The values being the copy-numbers.}

\item{Y}{\emph{matrix} A matrix of samples (x) by signatures (y). Values are signature exposures. \cr
\code{dim(Y)[1]} must equal \code{dim(X)[2]}}

\item{model_option}{(optional) \emph{character.} Which model to use. Options are \code{lm} or \code{beta}. \cr
For more modelling details see the \code{betareg} or \code{stats} R packages.}

\item{bin_indices}{(optional) \emph{numeric/integer vector.} A single integer or a vector of integers. \cr
Subsets the copy-number object to only run on the bins in these indices.}

\item{data_slot}{(optional) \emph{character.} If passing a QDNAseq object, the CN data slot to use. \cr
Common options are "copynumber" (corrected or uncorrected read counts), "segmented", or "calls".}

\item{run_parallel}{(optional) \emph{logical.} If TRUE run job in parallel. Will automatically use 1 less than total cores available.}

\item{cores}{(optional) \emph{integer.} Alternatively/in-addition, the user may choose the number of cores to use. \cr
Must be <= 1 less than the number available.}

\item{epsilon}{(optional) \emph{numeric.} Applies to the beta-modelling option only. \cr
For response values too near-to 0 or 1, and or subtract this amount +/- a small jitter.}
}
\value{
A list of lists. One entry for each bin. Each bin then has a list of: \cr
\code{model_predvals} \cr
\code{predictor_index}
}
\description{
Fit a model for each signature in Y given each bin of copy-numbers in CN dataset (X). \cr
The two model options are a linear model and a beta model.
}
