% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicimportance.R
\name{FitGIModelsStoCN}
\alias{FitGIModelsStoCN}
\title{Fit a model using CN-signature exposures to predict each bin in a CN-dataset}
\usage{
FitGIModelsStoCN(
  X,
  Y,
  bin_indices = NULL,
  data_slot = NULL,
  run_parallel = FALSE,
  cores = 1
)
}
\arguments{
\item{X}{\emph{matrix.} A matrix of samples (x) by signatures (y). Values are signature exposures. \cr
\code{dim(X)[1]} must equal \code{dim(Y)[2]}}

\item{Y}{\emph{S4/matrix.} A QDNAseq object that contains \code{copynumber}, and/or \code{segmented} assay slots. \cr
OR alternatively a matrix of bins (x) by samples (y) may be provided. The values being the copy-numbers.}

\item{bin_indices}{(optional) \emph{numeric/integer vector.} A single integer or a vector of integers. \cr
Subsets the copy-number object to only run on the bins in these indices.}

\item{data_slot}{(optional) \emph{character.} If passing a QDNAseq object, the CN data slot to use. \cr
Common options are "copynumber" (corrected or uncorrected read counts), "segmented", or "calls".}

\item{run_parallel}{(optional) \emph{logical.} If TRUE run job in parallel. Will automatically use 1 less than total cores available.}

\item{cores}{(optional) \emph{integer.} Alternatively/in-addition, the user may choose the number of cores to use. \cr
Must be <= 1 less than the number available.}
}
\value{
A list of lists. One entry for each bin. Each bin then has a list of: \cr
\code{model_predvals} \cr
\code{predictor_index}
}
\description{
Fit a model for each copy-number bin in Y given the signature exposures matrix X. \cr
The rows of X and columns of Y should match. \cr

Because signature exposures for a sample sum to 1, fitting a model as described encounters the issue of multi-collinearity.
FitGIModelsStoCN overcomes that issue by first performing lasso regression and then fitting a linear model (LM).
Lasso regression makes use of the \pkg{glmnet} package while the \pkg{stats} is called to fit each LM.

This function is a complement to \link{FitGIModelsCNtoS}.
}
