% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_levels.R
\name{GatherSegmentsByLevels}
\alias{GatherSegmentsByLevels}
\title{Merges segments by level.}
\usage{
GatherSegmentsByLevels(segments, granges_obj)
}
\arguments{
\item{segments}{A data frame containing segment data. For the first pass, these segments
are those with size >= 3mb. Their last column should be their level, and should be non-zero
(i.e. already determined)}

\item{granges_obj}{A GRanges object: used to obtain genomic data from regions we're merging.}
}
\description{
Iterates through given \code{segments} and merges by level or by chromosome arm. In other words,
we are keeping track of whenever we change chromosome arm or level.
}
