% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_helpers_gather_segments_ratiomedian.R
\name{GatherSegmentsByRatioMedianHelper}
\alias{GatherSegmentsByRatioMedianHelper}
\title{Merges segments that have the same ratio_median and are in the same chromosome arm.}
\usage{
GatherSegmentsByRatioMedianHelper(df)
}
\arguments{
\item{df}{The Data Frame with segment data: first column is chromosome number, second column is chromosome arm,
third column is start position, fourth is end position, and fifth is the ratio_median.}
}
\description{
Takes segment data (df) and gathers/merges the segments according to their ratio_median
or chromosome arm.
}
