% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureplotting.R
\name{GaussiansMixturePlot}
\alias{GaussiansMixturePlot}
\title{Plot a Mixture of Gaussians}
\usage{
GaussiansMixturePlot(
  signatures,
  components,
  sig_of_interest = 1,
  component = "segsize",
  inlay_flag = TRUE,
  threshold = 0.3
)
}
\arguments{
\item{signatures}{A dataframe with components along the y-axis and signatures along the x.}

\item{components}{A list of S4 objects belonging to the class 'flexmix'.}

\item{sig_of_interest}{(optional) A single integer or a vector of integers corresponding the signatures for which to make plots. \cr
In the absence of any value passed in for this parameter, plots for just the first signature will be returned.}

\item{component}{(optional) Which component to draw the gaussian curves for. (options: segsize, changepoint, copynumber)}

\item{inlay_flag}{(optional) Boolean flag to turn on/off the inlay plot.}
}
\value{
A ggplot.
}
\description{
This function plots the mixture models for a component that is composed of gaussians.
All mixture components are plotted, but those elevated for the indicated signature are shaded.
\cr
Additionally, an inlay plot is created out of the 'important' components if... \cr
\enumerate{
\item Some components have been 'squashed' down \cr
\item There are any. - i.e. at least 1 weight > 0.05 \cr
}
}
