% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainsignaturefunctions.R
\name{GenerateSampleByComponentMatrix}
\alias{GenerateSampleByComponentMatrix}
\title{Generate sum-of-posteriors probability matrix}
\usage{
GenerateSampleByComponentMatrix(
  CN_features,
  all_components = NULL,
  cores = 1,
  rowIter = 1000,
  subcores = 2
)
}
\arguments{
\item{CN_features}{A list. The output from either the \code{ExtractRelativeCopyNumberFeatures()} or \code{ExtractCopyNumberFeatures()} functions.}

\item{all_components}{A list of flexmix objects. One for each CN-feature.
Likely the output from \code{FitMixtureModels()}.}

\item{cores}{Integer. The number of cores to use for parallel processing.}

\item{rowIter}{Integer. Number of rows to ingest per iteration if using multiple cores. Otherwise ignored.}

\item{subcores}{Integer. In parallel mode each CN-feature will be run separately, this is the number for cores to allocate to each of these sub-jobs.}
}
\value{
A sum-of-posteriors probability matrix.
}
\description{
Given a set of extracted copy-number features and mixture models for each feature, generate a sum-of-posteriors matrix.
}
\details{
For each copy-number event for each sample the posterior probability of belonging to a component is computed.
These posterior event vectors are then summed resulting in a sum-of-posterior probabilities vector.
All sum-of-posterior vectors are combined into a single patient-by-component sum-of-posterior probabilities matrix.
}
