% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainsignaturefunctions.R
\name{GenerateSignatures}
\alias{GenerateSignatures}
\title{Create CN-Signatures}
\usage{
GenerateSignatures(
  sample_by_component,
  nsig,
  seed = 77777,
  nmfalg = "brunet",
  cores = 4
)
}
\arguments{
\item{sample_by_component}{Matrix. Expects a sum-of-posteriors probability matrix.}

\item{nsig}{Integer. The number of signatures. Likely decided by examining the plot from \code{ChooseNumberSignatures()}.}

\item{seed}{Integer. (flexmix param) The random seed to use while modelling.}

\item{nmfalg}{Character string. (NMF param) The specification of the NMF algorithm.}

\item{cores}{Integer. (NMF param) The number of cores to use when running NMF.}
}
\value{
An NMFfit object. Contains the two matrices post decomposition.
}
\description{
Create copy-number signatures by decomposing a sample-by-component matrix.
}
\details{
Given a sample-by-component sum-of-posteriors matrix and a matrix rank, decompose into two distinct matrices.
One matrix will be a sample-by-signature matrix and the other a signature-by-component matrix.
In this package, we refer to the signature-by-component matrix as a set of 'copy-number signatures'.
}
