% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarytable.R
\name{GenHumanReadableAcnProfile}
\alias{GenHumanReadableAcnProfile}
\title{Generate Human Readable Absolute Copy-Number Profiles}
\usage{
GenHumanReadableAcnProfile(object, save_path)
}
\arguments{
\item{object}{S4 copy-number object - QDNAseq or CGHcall object}

\item{save_path}{A string. A path (directory) to where segment tables should be saved. ex. '~/Documents/test_project'}
}
\value{
A table. Collapsed Segment tables in long format (by sample id) ready to be written out to a table file (ex. tsv, csv).
}
\description{
This function takes as input a QDNAseq or CGHcall copy-number object and gives back Cytoband .tsv tables from absolute copy-number calls (in a human readable format).
}
