% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GenHumanReadableRcnProfile}
\alias{GenHumanReadableRcnProfile}
\title{Generate Human Readable Relative Copy-Number Profiles}
\usage{
GenHumanReadableRcnProfile(object, binsize, ref_genome, save_dir = FALSE)
}
\arguments{
\item{object}{S4 copy-number object - QDNAseq or CGHcall object}

\item{binsize}{The binsize used in the copy number object. ex. '30kb', '100kb'}

\item{ref_genome}{One of the common reference genomes: ex. 'hg19', 'mm10', or 'hg38'}

\item{save_dir}{(optional) The directory where the tables should be saved. ex. '~/Documents/test_project'}
}
\value{
Segment tables in long format (by sample id) ready to be written out to a table file (ex. tsv, csv).
}
\description{
This function takes as input a QDNAseq or CGHcall copy-number object and gives back a long-format table with several useful columns.
These columns include; 'sample', 'chromosome', 'start', 'end', 'gain_probability',
'loss_probability', 'relative_copy_number', 'bin_count', 'sum_of_bin_lengths',
'cytobands', 'coordinates', and 'size'
}
