% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_helpers_find_threshold.R
\name{GetCriticalPoints}
\alias{GetCriticalPoints}
\title{Finds critical points (minima or maxima) of the differences between
segments ratios.}
\usage{
GetCriticalPoints(batch_ratio_differences, crit_point)
}
\arguments{
\item{crit_point}{A string. The critical point to look for: either 'max' or 'min'.}

\item{ratio_differences}{A Data Frame. The estimated density values for the given sampled batch}
}
\description{
GetCriticalPoints takes the estimated density values (via KDE) from a batch of samples
of segment ratio differences (S_i - S_j), a critical point (min/max), and returns a
list of the densities where the critical point is found.

If crit_point == 'max', then densities > 0
If crit_point == 'min', then densities < 0
}
