% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_small_segments.R
\name{GetGRangesObject}
\alias{GetGRangesObject}
\title{Returns a GRanges object used to obtain data from specified genomic regions.}
\usage{
GetGRangesObject(bam_ratios_frame)
}
\arguments{
\item{bam_ratios_frame}{A data frame: the cleaned-up version of the raw bam ratios file.}
}
\description{
As above. Usually used to get data from a resulting merge, or from missing segments
in between segments we already have data for.
}
