% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_call_lga.R
\name{GetHRDStatus}
\alias{GetHRDStatus}
\title{Returns the HRD status of a sample}
\usage{
GetHRDStatus(lga_calls)
}
\arguments{
\item{lga_calls}{The table of LGA counts by size, as returned by CallLGA.}
}
\value{
The HRD status of the sample: TRUE if the number of LGAs >= 10 Mb exceeds 20; FALSE otherwise.
}
\description{
Using the output of CallLGA, we determine the HRD status of a sample based on the number of LGAs >= 10 Mb in size.
}
