% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_levels.R
\name{GetLargeSegments}
\alias{GetLargeSegments}
\title{Returns segments with a size >= 3Mb.}
\usage{
GetLargeSegments(segments)
}
\arguments{
\item{segments}{}
}
\description{
Returns a data frame with segments whose size is larger than 3 mega bases.
}
