% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_call_lga.R
\name{GetLGAOfSize}
\alias{GetLGAOfSize}
\title{Returns the segments that are LGAs of the given size}
\usage{
GetLGAOfSize(threshold, size_lga, segments)
}
\arguments{
\item{threshold}{A float. Threshold in ratio_median difference previously estimated.}

\item{size_lga}{An integer: size of the LGA to look for.}

\item{segments}{A data frame: segment data}
}
\description{
With help of DetermineNumberOfLGAs, keeps segment data for LGA size of \code{size_lga}
}
