% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplequality.R
\name{GetSampleQualityDecision}
\alias{GetSampleQualityDecision}
\title{Calculate and make a quality call for relative copy number profiles}
\usage{
GetSampleQualityDecision(x, metric = "quantile", cutoff = 0.95)
}
\arguments{
\item{x}{\emph{QDNASeq object} containing the relative copy number calls as well as the segmented relative copy number calls}
}
\value{
A dataframe of sample quality decisions ("Low" or "High" quality sample)
}
\description{
Function to classify the quality of relative copy number profile calls from QDNAseq or WiseCondorX
}
\details{
Expects the input of a QDNAseq object containing relative copy-number calls (such as from QDNASeq or WiseCondorX). \cr
Classifies relative copy number profiles as "High" or "Low" quality.
}
