% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplequality.R
\name{GetSegCounts}
\alias{GetSegCounts}
\title{Calculate segment sizes}
\usage{
GetSegCounts(x)
}
\arguments{
\item{x}{Dataframe containing the following columns: chromosome, start, stop, segmented, copy_number and new_segment (indicator variable for the segment the copy_number call belongs to)}
}
\description{
Calculate segment sizes
}
