% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_call_lga.R
\name{GetSegmentationBeforeLGACall}
\alias{GetSegmentationBeforeLGACall}
\title{Final segmentation before LGA call}
\usage{
GetSegmentationBeforeLGACall(
  segments,
  bam_ratios_frame,
  granges_obj,
  second_round
)
}
\arguments{
\item{segments}{A data frame of segments with size > 3Mb. Ideally this
frame has already gone through merging, small segments reinsertion, etc.}

\item{bam_ratios_frame}{A data frame. For first pass: the cleaned-up version of the initial
ratio file tsv. For second pass: this clean-up segments but merged by ratio_median.}

\item{granges_obj}{A GRanges object: used to obtain genomic data from regions we're merging.}

\item{second_round}{A boolean: FALSE if first pass, TRUE if second pass.}
}
\description{
Merges the initial \code{bam_ratios_frame} by segments in the same chromosome,
and then assigns start/end pos to segments in \code{segments}, whenever
we change Chr_N in the latter.
}
