% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_helpers_call_lga.R
\name{GetSegmentID}
\alias{GetSegmentID}
\title{Re-assigns level column values for all segments.}
\usage{
GetSegmentID(threshold, segments)
}
\arguments{
\item{threshold}{A float: threshold for ratio_median difference previously estimated.}

\item{segments}{A data frame: segment data.}
}
\description{
Resets level to 0 for all segments, except for the first segment which is assigned 1.
Then, iterates through the rest:
For a segment 'i', if it is in the same chromosome arm as previous segment 'i-1' and
their ratio median difference is below the threshold, then segment 'i' gets 'i-1''s level.
Otherwise, it is 'i-1''s level + 1.
}
