% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_small_segments.R
\name{GetSmallSegments}
\alias{GetSmallSegments}
\title{Returns the segments with 3Mb >= size >= 0.1Mb}
\usage{
GetSmallSegments(segments)
}
\arguments{
\item{segments}{A data frame containing the segment data.}
}
\description{
Filters the segments and returns only the small ones
}
