% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_helpers_small_segments.R
\name{InitalizeSmallSegments}
\alias{InitalizeSmallSegments}
\title{Initializes the process of re-inserting the small segments in between the large ones.}
\usage{
InitalizeSmallSegments(large_segments, small_segments, threshold, granges_obj)
}
\arguments{
\item{large_segments}{A data frame. Segments with size >= Mb.}

\item{small_segments}{A data frame. Segments with 3Mb >= size >= 0.1Mb.}

\item{threshold}{A float: the estimated threshold for ratio_median difference via KDE. Used to determine
whether we insert the small segment or not.}

\item{granges_obj}{A GRanges object: is used as genomic reference to check whenever we have
an overlap of segments and get the data (specifically the ratio_median) of this overlap.}
}
\description{
Essentially deals with the FIRST large segment. We add small segments from \code{small_segments}
segments to \code{large_segments} that:
\enumerate{
\item End before the first large segment and
\item Overlap with the first large segment
We stop the process whenever we've changed chromosome arms between the small segment and the large segment
}
}
