% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_small_segments.R
\name{InsertSmallSegments}
\alias{InsertSmallSegments}
\title{Adds small segments in \code{small_segments} that meet certain criteria (position wise w.r.t large segments)
into the \code{large_segments} data frame. Merges if necessary (i.e. if ratio_median difference < threshold).}
\usage{
InsertSmallSegments(large_segments, small_segments, threshold, granges_obj)
}
\arguments{
\item{large_segments}{A data frame. Segments with size >= 3Mb.}

\item{small_segments}{A data frame. Segments with 3Mb > size >= 0.1Mb.}

\item{threshold}{A float: the estimated threshold for ratio_median difference via KDE.}

\item{granges_obj}{A GRanges object: is used as reference to check whenever we have
an overlap of segments and get the ratio_median of this overlap.}
}
\description{
Does a lot of things. In InitializeSmallSegments, we insert all the small segments that come before the first
large segment in \code{large_segments} (hence initializing). In FinalizeSmallSegments, we insert all the remaining small segments,
according to 6 different cases, which are listed in the FinalizeSmallSegments documentation. In the end, returns the updated
\code{large_segments} with small segments appropriately re-inserted.
}
