% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_small_segments.R
\name{LargeMissingChrArms}
\alias{LargeMissingChrArms}
\title{Deals with missing chromosome arm values in the given data frame of large segments.}
\usage{
LargeMissingChrArms(large_segments, small_segments)
}
\arguments{
\item{large_segments}{A data frame. Segments with size >= Mb.}

\item{small_segments}{A data frame. Segments with 3Mb >= size >= 0.1Mb.}
}
\description{
Finds the missing chromosome arms and reinserts them into \code{large_segments}. Specifically,
the arms that \code{large_segments} doesn't have, but \code{small_segments} does. For each of the missing
values, we iterate through the large segments, and re-insert the small segments wherever
appropriate.

For example (first case in while loop), if the missing_chr_arm value is greater than the largest chr_arm
in the arm column (i.e., max(large_segments\link{,3})), then we should bind the small segments
to the right of the large segments frame, since there are no more segments after
this largest chr_arm. Note that we only bind the small segments within this missing_chr_arm.

Otherwise, missing_chr_arm <= than the largest chr_arm. In this case we care about which large
segment we're looping over (the value of \code{i}). If i = 1, i.e. the first segment, then we bind
the small segments to the left of the large segments: this is trivial. Otherwise, we fit the
small segments in between \code{i-1} and \code{i}.
}
