% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarytable.R
\name{MakeGISummaryTable}
\alias{MakeGISummaryTable}
\title{Makes a summary table of per-gene coefficients and pvals from GI output}
\usage{
MakeGISummaryTable(
  GI_vectors,
  ref_genome = NULL,
  coef_threshold = 0.1,
  pval_threshold = 0.05,
  rm_blacklist = TRUE,
  edb = EnsDb.Hsapiens.v75::EnsDb.Hsapiens.v75
)
}
\arguments{
\item{GI_vectors}{\emph{list.} A list of vectors (Ex. the output from \link{FitGIModelsCNtoS}). \cr
Two vectors (pvals and coefficients) for each signature. \cr
One additional vector of the genomic bins for which the regression values correspond.
All vectors must have the same length.}

\item{ref_genome}{\emph{character.} The reference genome used in creating this dataset. ex. "hg19" \cr
Make sure this char vector matches the database provided to the 'edb' parameter.}

\item{coef_threshold}{\emph{numeric.} Don't consider any bins where the coef falls below this value.}

\item{pval_threshold}{\emph{numeric.} Don't consider any bins where the p-value falls above this value.}

\item{rm_blacklist}{\emph{logical.} If TRUE, remove the blacklist regions for this genome (low mappability etc.).
If FALSE, then don't.}

\item{edb}{\emph{Formal class EnsDb.} This parameter expects an ensembl DB object for the corresponding genome provided to the 'ref_genome' param. \cr
Ex. EnsDb.Hsapiens.v75::EnsDb.Hsapiens.v75 or EnsDb.Mmusculus.v79::EnsDb.Mmusculus.v79}
}
\value{
A data.table. One row per gene composed of the median calculated GI values across the corresponding bins.
}
\description{
MakeGISummaryTable takes as input a list of vectors (Ex. the output from \link{FitGIModelsCNtoS}). \cr
For each provided bin that passes the coefficient and p-value thresholds the function finds all genes that overlap the region.
It then builds an output table composed of the median calculated GI values per-gene.
The output tables for each input signature are all stacked.
}
