% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicimportance.R
\name{MakeGIVectors}
\alias{MakeGIVectors}
\title{Reformat genomic importance modelling output}
\usage{
MakeGIVectors(results)
}
\arguments{
\item{results}{\emph{list of lists.} The output from FitGIModelsCNtoS.}
}
\value{
A list of vectors. Two vectors (pvals and coefficients) for each signature. \cr
All vectors have the same length as the provided number of bins to FitGIModelsCNtoS.
}
\description{
Reformat genomic importance modelling output into a list of vectors.
}
