% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplequality.R
\name{MedSegVar}
\alias{MedSegVar}
\title{Calculate median segment-level variance per sample}
\usage{
MedSegVar(x)
}
\arguments{
\item{x}{Dataframe containing the following columns: chromosome, start, stop, segmented, and copy_number and new_segment (indicator variable for the segment the copy_number call belongs to)}
}
\description{
Calculate median segment-level variances per sample (median absolute deviance)
}
