% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_helpers_small_segments.R
\name{MergeSegmentsFiveLargeSmall}
\alias{MergeSegmentsFiveLargeSmall}
\title{Deals with merging large segment to small segment.}
\usage{
MergeSegmentsFiveLargeSmall(
  granges_obj,
  small_segment,
  large_segments,
  j,
  end_of_file
)
}
\arguments{
\item{granges_obj}{A GRanges object: is used as reference to obtain true genomic data from specified regions.}

\item{small_segment}{An array: the given small segment that might be merged into the \code{large_segments} frame.}

\item{large_segments}{A data frame: segment data. Initially contains only segments with size >= 3Mb, but as we merge
small segments into it, smaller segments are progressively added.}

\item{j}{An integer/index: the position of the large segment we're comparing with small_segment, within \code{large_segments}.}

\item{end_of_file}{A boolean: whether there's any large segments after \code{j}.}
}
\description{
Procedure is overall similar to the previous cases. For case five, we have two options when merging:
\itemize{
\item Merge large and small
\item Merge small and next large
This helper deals with the first option.
Note that there is no 'below_thr' parameter for these two helpers, since we already know we're going to merge
when they're called in FinalizeSmallSegments.
}
}
