% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_helpers_find_threshold.R
\name{NullListOfThresholds}
\alias{NullListOfThresholds}
\title{This function is ran in case that no possible thresholds were found.}
\usage{
NullListOfThresholds(all_ratio_differences)
}
\arguments{
\item{all_ratio_differenves}{An array: the ratio differences for all pairs of segments.}
}
\description{
NullListOfThresholds returns a threshold in the case where list_of_possible_threshold was null.
It finds the first local minima, but also finds the inflection points of the density function, and
returns as threshold the min of these: min(local_minia, inflection_point)
}
