% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot,QDNAseqSignals,missing-method}
\alias{plot,QDNAseqSignals,missing-method}
\title{Plot Relative Copy Numbers}
\usage{
\S4method{plot}{QDNAseqSignals,missing}(
  x,
  y,
  main = NULL,
  includeReadCounts = TRUE,
  logTransform = TRUE,
  scale = TRUE,
  sdFUN = "sdDiffTrim",
  delcol = getOption("QDNAseq::delcol", "darkred"),
  losscol = getOption("QDNAseq::losscol", "red"),
  gaincol = getOption("QDNAseq::gaincol", "blue"),
  ampcol = getOption("QDNAseq::ampcol", "darkblue"),
  pointcol = getOption("QDNAseq::pointcol", "black"),
  segcol = getOption("QDNAseq::segcol", "chocolate"),
  misscol = getOption("QDNAseq::misscol", NA),
  pointpch = getOption("QDNAseq::pointpch", 1L),
  pointcex = getOption("QDNAseq::pointcex", 0.1),
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  xaxt = "s",
  yaxp = NULL,
  showDataPoints = TRUE,
  showSD = TRUE,
  doSegments = TRUE,
  doCalls = TRUE,
  ...,
  verbose = getOption("QDNAseq::verbose", TRUE)
)
}
\description{
Just an updated version of the \pkg{QDNAseq} package's plot() method.
This version just allows for relative copy number calls from WiseCondorX too.
All else remains the same.
See the <\code{\link[QDNAseq:plot]{QDNAseq plotting}}> method for details.
}
\details{
Supply a QDNAseq object.
}
