% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tallysignaturefunctions.R
\name{PlotCNTallySig}
\alias{PlotCNTallySig}
\title{Plot Copy Number Tally Signature Contributions}
\usage{
PlotCNTallySig(signatures, extra_features = FALSE)
}
\arguments{
\item{signatures}{An NMF object containing the copy number signature data, typically the output of a
non-negative matrix factorization (NMF) process. The function extracts the basis matrix from the NMF object.}

\item{extra_features}{A logical value indicating whether extra features are included (default: FALSE).}
}
\value{
A \code{ggplot2} plot object visualizing the contribution of each signature to copy number components.
The plot consists of faceted bar charts, where each facet represents a signature and the x-axis
corresponds to component bins (e.g., specific copy number ranges).
}
\description{
This function visualizes copy number signature contributions across different components.
It generates a bar plot where the x-axis represents component bins (e.g., copy number categories),
and the y-axis shows the normalized contributions of each signature to the components.
}
