% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_plot_segments.R
\name{PlotSegmentChanges}
\alias{PlotSegmentChanges}
\title{A plot illustrating the merging of segments by the SHRD algorithm.}
\usage{
PlotSegmentChanges(
  segments_initial,
  segments_pass_one,
  segments_pass_two,
  sample,
  segment_line_size = 4
)
}
\arguments{
\item{segments_initial}{A data frame of the initial gathered segments.}

\item{segments_pass_one}{A data frame of the gathered segments, after processing and merging in the first pass.}

\item{segments_pass_two}{A data frame of the gathered segments, after processing and merging in the second pass.}

\item{sample}{The sample name.}

\item{segment_line_size}{The line width for the initial segments. Defaults to 4; first-pass & second-pass segments will be 1/2 and 1/4 this width, respectively.}
}
\value{
A plot illustrating the initial, first-pass, and second-pass (final) segments.
}
\description{
A plot illustrating the merging of segments by the SHRD algorithm.
}
