% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_plot_segments.R
\name{PlotSegments}
\alias{PlotSegments}
\title{Plots segments in the given data frame.}
\usage{
PlotSegments(
  gathered_by_ratio_median,
  bam_ratios_frame,
  segments,
  chr_mid_positions
)
}
\arguments{
\item{gathered_by_ratio_median}{A data frame: the segments merged by the same
ratio_median and/or chromosome arm. Most importantly, the frame must have already been
prepped for initialization of Levels in \code{PrepForInitialization()}. This is also
marked as 'Graph 1' in \code{stitching_functions.R}. The frame is used to find the max and min
values of the ratio_median, which corresponds to the y-axis of the graph.}

\item{bam_ratios_frame}{A data frame: the cleaned-up version of the bam_ratios.txt file.}

\item{segments}{A data frame: segment data. Usually is Graph 5 in the stitching script.}

\item{chr_mid_positions}{An array: the middle positions for each chromosomes from 1 to 23, in that order.}
}
\description{
Plots segments in their respective chromosomes with the genome's middle positions
from each chromosome, obtained in \code{hg19_segments.R}
}
