% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureplotting.R
\name{PoissonsMixturePlot}
\alias{PoissonsMixturePlot}
\title{Plot a Mixture of Poissons}
\usage{
PoissonsMixturePlot(
  signatures,
  components,
  sig_of_interest = 1,
  component = "bp10MB",
  threshold = 0.3
)
}
\arguments{
\item{signatures}{A dataframe with components along the y-axis and signatures along the x.}

\item{components}{A list of S4 objects belonging to the class 'flexmix'.}

\item{sig_of_interest}{(optional) A single integer or a vector of integers corresponding the signatures for which to make plots. \cr
In the absence of any value passed in for this parameter, plots for just the first signature will be returned.}

\item{component}{(optional) Which component to draw the poisson 'curves' for. (options: bp10MB, osCN, bpchrarm)}
}
\value{
A ggplot.
}
\description{
This function plots the mixture models for a component that is composed of poisson distributions.
All mixture components are plotted, but those elevated for the specified signature are shaded.
}
