% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_helpers_find_threshold.R
\name{PrepFirstRound}
\alias{PrepFirstRound}
\title{Preps the segment data frame for use in the first round of threshold estimation.}
\usage{
PrepFirstRound(granges_obj, segments)
}
\arguments{
\item{granges_obj}{A GRanges object: used to fine genomic data from the specified regions.}

\item{segmenta}{A data frame: segment data.}
}
\description{
Only keeps segments whose size is >= 3Mb. First iterates through segments and assigns genomic
data from GRanges to this segment (specifically their ratio_median)
}
