% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_levels.R
\name{PrepForLevelsInitialization}
\alias{PrepForLevelsInitialization}
\title{Calls some helpers to do something to gathered by ratio medians.... gotta check this
Also adds a Level and an Index column to the data frame}
\usage{
PrepForLevelsInitialization(segments, granges_obj)
}
\arguments{
\item{segments}{The segments gathered_by_ratio_medians}

\item{granges_obj}{A granges object containing the clean ratios file after CleanBamRatios()}
}
\description{
Adds couple of columns to the given segment dataframe + other processing things
}
