% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_helpers_find_threshold.R
\name{PrepSecondRound}
\alias{PrepSecondRound}
\title{Preps the segment data frame for use in the second round of threshold estimation.}
\usage{
PrepSecondRound(segments)
}
\arguments{
\item{segments}{A data frame: segment data}
}
\description{
Removes columns 1, 7, and 8 from \code{segments}, also readjusts segment sizes.
}
