% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{QualityPlot}
\alias{QualityPlot}
\title{Plot sample quality according to a specified metric.}
\usage{
QualityPlot(sample_quality_df, metric = "quantile")
}
\arguments{
\item{sample_quality_df}{The sample quality data-frame returned by GetSampleQualityDecision().}

\item{metric}{The quality metric passed to GetSampleQualityDecision(). One of quanitle, or DecisionTree.}
}
\value{
A ggplot object of the quality plot.
}
\description{
Plot sample quality according to a specified metric.
}
