% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainsignaturefunctions.R
\name{QuantifySignatures}
\alias{QuantifySignatures}
\title{Quantify Signature Exposures}
\usage{
QuantifySignatures(sample_by_component, component_by_signature = NULL)
}
\arguments{
\item{sample_by_component}{Matrix. Provide a matrix of posteriors for each sample to each component in the model used.}

\item{component_by_signature}{String or object of type \code{NMFfitX1}. If a string, the function expects a path to CN-Signatures (an \code{NMFfitX1} object) stored as an rds.}
}
\value{
A matrix of the exposures for each sample to each signature.
}
\description{
Given a set of \cr
\enumerate{
\item signatures and \cr
\item posteriors for each extracted CN-feature across samples to a set of component models \cr
calculates signature exposures.  \cr
}

Said more concretely, given a sample-by-component matrix and a component-by-signature matrix, finds the linear combination decomposition (LCD).
See the YAPSA::LCD function for details.
}
