% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_helpers_small_segments.R
\name{RatioDiffBelowThreshold}
\alias{RatioDiffBelowThreshold}
\title{Determines if absolute ratio_median difference is less than or equal to the given threshold.}
\usage{
RatioDiffBelowThreshold(large_segment, small_segment, threshold)
}
\arguments{
\item{large_segment}{An array: a single segment, doesn't necessarily have to be a 'large' segment.
This name was chosen because most of the times we're comparing a large segment in \code{large_segments} with
a small segment in \code{small_segments}}

\item{small_segment}{Same as above.}

\item{threshold}{A float: the threshold for ratio_median difference previously estimated.}
}
\description{
Returns TRUE if large_segment and small_segment's ratio_median difference is <= threshold. FALSE otherwise.
}
