% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{RCNDiversityPlot}
\alias{RCNDiversityPlot}
\title{Plot relative CN diversity heatmap}
\usage{
RCNDiversityPlot(
  qdnaseq_obj,
  order_by = NULL,
  cluster = TRUE,
  subset = NULL,
  Xchr = FALSE,
  limits = c(-1.5, 1.5),
  breaks = c(2.5, 1, 0.5, 0, -0.5, -1, -2.5),
  genome = "hg19"
)
}
\arguments{
\item{order_by}{optional \emph{character vector} containing the order the samples in the heatmap should be in}

\item{cluster}{optional \emph{logical} Do you want to do a row-wise sort of the  heatmap (sort the samples)}

\item{subset}{optional \emph{character vector} A subset of samples that you'd like to plot.}

\item{limits}{\emph{vector} containing the limits of your colour gradient. Passed to scale_fill_gradientn}

\item{breaks}{\emph{vector} containing the limits of your colour gradient. Passed to scale_fill_gradientn}

\item{genome}{reference genome build}

\item{x}{\emph{dataframe} containing the copy number segments in long format: columns should be start, sample id (named sample) and segmented}
}
\value{
Heatmap of relative copy number calls
}
\description{
Function to plot the copy number profile from relative copy numbers
}
\details{
Visualizing relative copy number calls from WiseCondorX and QDNASeq as heatmaps
}
