% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{rCNplotProfile}
\alias{rCNplotProfile}
\title{Plot relative CN diversity heatmap}
\usage{
rCNplotProfile(
  x,
  order_by = NULL,
  cluster = TRUE,
  subset = NULL,
  limits = c(-1, 1),
  breaks = c(-2.5, -1, -0.75, 0, 0.75, 1)
)
}
\arguments{
\item{x}{\emph{dataframe} containing the copy number segments in long format: columns should be start, sample id (named sample) and segmented}

\item{order_by}{optional \emph{character vector} to plot a subset of samples}

\item{limits}{\emph{vector} containing the limits of your colour gradient. Passed to scale_fill_gradientn}

\item{breaks}{\emph{vector} containing the limits of your colour gradient. Passed to scale_fill_gradientn}
}
\value{
Heatmap of relative copy number calls
}
\description{
Function to plot the copy number profile from relative copy numbers
}
\details{
Visualizing relative copy number calls from WiseCondorX and QDNASeq as heatmaps
}
