% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copynumberscaling.R
\name{ReplaceQDNAseqAssaySlots}
\alias{ReplaceQDNAseqAssaySlots}
\title{Replacethe QDNAseq Assay Slots}
\usage{
ReplaceQDNAseqAssaySlots(cnobj, new_cns, new_segs)
}
\description{
Replace the copynumber and/or segment QDNAseq slots with new matrices
}
