% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_runshallowhrd.R
\name{RunShallowHRDFromQDNA}
\alias{RunShallowHRDFromQDNA}
\title{Run ShallowHRD on a QDNAseq object.}
\usage{
RunShallowHRDFromQDNA(
  qdna_obj,
  include_chr_X = FALSE,
  num_simulations = 100000,
  shrd_save_path = FALSE,
  plot = FALSE,
  seed = 1337,
  cores = 1
)
}
\arguments{
\item{qdna_obj}{An S4 object of type QDNAseqCopyNumbers. This object must contain a copynumber slot and a segmented slot.}

\item{include_chr_X}{Whether or not to include the X-chromosome when running ShallowHRD.}

\item{num_simulations}{The number of simulations to use during the FindThreshold() step.}

\item{shrd_save_path}{An optional path to save the ShallowHRD results to.}

\item{plot}{Whether or not to return a plot showing how segments are merged as the algorithm runs.}

\item{seed}{A seed to use for PRNG-dependent functions. Ensures reproduciblity between runs.}

\item{cores}{The number of cores to use for running samples in parallel. If set to 1, no parallel processing will be used.}
}
\value{
A list of ShallowHRD results, with one entry per sample.
}
\description{
This function takes a QDNAseqCopyNumbers object, extracts the binned and segmented copy-number data, and runs ShallowHRD on each sample.
}
