% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplequality.R
\name{SampleGrouping}
\alias{SampleGrouping}
\title{Extract sample grouping}
\usage{
SampleGrouping(x)
}
\arguments{
\item{x}{param_dat containing the following columns: sample_id and sample quality parameters}
}
\description{
Extract sample grouping
}
