% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureplotting.R
\name{SEAlluvialPlot}
\alias{SEAlluvialPlot}
\title{Make an Alluvial Plot based on maximum signature exposure}
\usage{
SEAlluvialPlot(exposA, exposB, orderA = NULL, orderB = NULL)
}
\arguments{
\item{exposA}{A matrix of signature exposures. Signatures in rows, samples in columns. Likely the output from \code{CallSignatureExposures()}.}

\item{exposB}{A matrix of signature exposures. Signatures in rows, samples in columns.}

\item{orderA}{(Optional) A character vector of signature names that will enforce an order. \cr
Signature set A is assigned names S.A1 -> S.A#, and B S.B1 -> S.B# \cr
ex. c('S.B2', 'S.B5', 'S.B3', 'S.B4', 'S.B1', 'S.B6')}

\item{orderB}{(Optional) A character vector of signature names that will enforce an order. \cr
Signature set A is assigned names S.A1 -> S.A#, and B S.B1 -> S.B# \cr
ex. c('S.B2', 'S.B5', 'S.B3', 'S.B4', 'S.B1', 'S.B6')
Must be for the same set of same set of samples as provided to \code{exposA}. Likely the output from \code{CallSignatureExposures()}.}
}
\value{
A \emph{list}. A ggplot and its corresponding data matrix.
}
\description{
SEAlluvialPlot builds a ggplot using the easyalluvial R package.
For two sets of exposures on the same set of samples it visualizes how the maximum exposure for one set of signatures is related to the other.
}
