% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrd_helpers_call_lga.R
\name{ShrinkReprTMP}
\alias{ShrinkReprTMP}
\title{(Kind of) merges segments that lie on the same level.}
\usage{
ShrinkReprTMP(segments, granges_obj)
}
\arguments{
\item{segments}{A data frame: segment data.}

\item{granges_obj}{A GRanges object.}
}
\description{
Iterates through segments in \code{segments}, and if two segments have the same level,
then they are essentially merged with help of \code{granges_obj}. Actual procedure described
in for-loop below.
Removes the segments that were not merged.
}
