% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureplotting.R
\name{SignatureExposuresPlot}
\alias{SignatureExposuresPlot}
\title{Create Heatmap of Signature Exposures}
\usage{
SignatureExposuresPlot(
  signatures,
  order = FALSE,
  transpose = FALSE,
  colour_scheme = "D",
  colour_dir = 1,
  cm_begin = 0,
  addtitle = NULL,
  obj_name = "sig_exposures_obj"
)
}
\arguments{
\item{signatures}{Dataframe. Expects a dataframe of signature exposures (rows) by samples (columns).}

\item{order}{(optional) Character vector. Defines the order in which samples will be plotted. \cr
This is particularly useful when plotting more than 1 heatmap next to one another. \cr
Allows samples to line-up horizontally. \cr
Example: \cr
\verb{c("CC-CHM-1341", "CC-CHM-1347", "CC-CHM-1355", CC-CHM-1361", "CC-HAM-0369", "CC-HAM-0374", "CC-HAM-0379", "CC-HAM-0383", "CC-HAM-0385")}}

\item{transpose}{(optional) Logical. If set to TRUE the function returns the order in which samples were plotted.}

\item{colour_scheme}{(optional) Character. Value passed to the \code{viridis::scale_fill_viridis} option parameter. ex. 'A' or 'B'}

\item{colour_dir}{(optional) Either 1 OR -1. Value passed to the \code{viridis::scale_fill_viridis} direction parameter.}

\item{cm_begin}{(optional) Float. Value within range \link{0,1}. Value passed to the \code{viridis::scale_fill_viridis} direction parameter.}

\item{obj_name}{(optional) String. Adds a tag to the end of the filename if saving the image. \cr
Only used if the save_path parameter is also set.}

\item{title}{(optional) String or TRUE. Pass in a title for the plot or set to TRUE and one will be generated.}
}
\value{
A list of a ggplot2 object and vector of the ordered sample names.
}
\description{
Converts signature-per-sample data to a heatmap, optionally saves it as a png, and returns the ggplot.
Samples are sorted for display based on their maximum signature exposure.
Heatmap is plotted in the viridis colour-scheme.
}
