% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureplotting.R
\name{StackedExposuresPlot}
\alias{StackedExposuresPlot}
\title{Create a Stacked Bar-plot of Signature Exposures}
\usage{
StackedExposuresPlot(
  input_matrix,
  user_colours = NULL,
  orderbyrow = 1,
  orderbyname = NULL,
  do_transpose = FALSE
)
}
\arguments{
\item{input_matrix}{A dataframe/matrix/datatable with signatures in the rows and samples in the columns..}

\item{user_colours}{(optional) A vector of colours for the signatures, one per signature.}

\item{orderbyrow}{(optional) An integer. Single integer corresponding to the row by which to sort.}

\item{orderbyname}{(optional) NULL or a character vector of sample names.}

\item{do_transpose}{(optional) Logical. If TRUE then flip bars horizontal.}
}
\value{
A ggplot2 object.
}
\description{
Converts signature-per-sample data into a ggplot2 stacked bar plot and returns that object.
Samples are sorted by signature exposure (default is the first row).
This can be changed to a different row, or a char vector or sample names can be supplied for a custom ordering.
The turbo viridis colour-scheme is used.
}
