% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copynumberscaling.R
\name{TumourFraction}
\alias{TumourFraction}
\title{Compute tumour DNA fraction for the given absolute copy number and
cellularity}
\usage{
TumourFraction(absolute_copy_number, cellularity)
}
\arguments{
\item{absolute_copy_number}{the absolute copy number(s) (a numeric value).}

\item{cellularity}{the cellularity, i.e. the fraction of cells that are from
the tumour.}
}
\value{
the fraction of DNA that originates from tumour cells.
}
\description{
Compute the tumour DNA fraction for the given absolute copy number(s) in the
tumour and the cellularity, i.e. the fraction of all cells that are tumour
cells.
Copied from rascal.
}
\examples{
tumour_fraction(3, 0.82)
}
