% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureplotting.R
\name{TwoFeatureScatterPlot}
\alias{TwoFeatureScatterPlot}
\title{Make a two-way scatterplot of two CN-features}
\usage{
TwoFeatureScatterPlot(featA, featB)
}
\description{
This function makes a two-way scatterplot of two CN-features.
It then draws contour lines for density.
}
