% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureplotting.R
\name{WassDistancePlot}
\alias{WassDistancePlot}
\title{Plot a Heatmap of Component-wise Wasserstein Distances}
\usage{
WassDistancePlot(cm_a, cm_b, component)
}
\arguments{
\item{cm_a}{A list of S4 objects each of class 'flexmix'. Likely the output from the FitMixtureModels function.}

\item{cm_b}{A list of S4 objects each of class 'flexmix'. Likely the output from the FitMixtureModels function.}

\item{component}{Which CN-feature models to compare to one another. (options: segsize, copynumber, bp10MB, osCN, bpchrarm)}
}
\value{
A \emph{list}. One ggplot and its corresponding matrix.
}
\description{
This function calculates and builds a heatmap of component-wise wasserstein distances between two mixture models. \cr
Also known as calculating the 'Earth Mover's Distance'.
A CN-feature is indicated by the user using the "component" parameter for this function.
The mixture models for this feature in each set of component models are then pulled out and compared.
In short, for each component in each mixture model indicated, a distance will be calculated.
}
