% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copynumberscaling.R
\name{CalculateACNs}
\alias{CalculateACNs}
\title{Calculate Absolute Copy Numbers}
\usage{
CalculateACNs(
  cnobj,
  acnmethod,
  rascal_sols = NULL,
  variants = NULL,
  acn_save_path = FALSE,
  return_sols = FALSE,
  return_S4 = FALSE,
  distance_decimal_places = 7
)
}
\arguments{
\item{cnobj}{An S4 object of type QDNAseqCopyNumbers. This object must contain a copynumber slot and a segmented slot.}

\item{acnmethod}{The method by which to calculate ACNs. Can be one of: \cr
"maxvaf" - Calculate ACNs assuming the maximum discovered VAF for the sample is an appropriate representation for the tumour fraction. \cr
char vector - Same as above but rather than using the max vaf provide a character vector of the genes from which to pull VAFs.
The genes are assumed to be in order of decreasing precedence. ex. c('TP53', 'KRAS', 'PTEN') \cr
"mad" - Calculate ACNs using the mean absolute difference (MAD) column from the solutions table. \cr
If using variant allele frequencies from targeted panel sequencing or some other technology: \cr
\itemize{
\item The variants must must be in a datatable/dataframe. \cr
\item Required columns: sample_id, chromosome, start, end, gene_name, ref, alt, vaf.
\item Each row of said table must correspond to a unique variant. \cr
\item Each variant must have an associated variant allele frequency. \cr
\item Each row must also be associated with a specific sample. \cr
A dataframe of known ploidies per sample - Calculate ACNs at each bin given we already know the ploidy of the samples. \cr
}}

\item{rascal_sols}{A tsv or dataframe of the calculated rascal solutions.}

\item{variants}{(optional) Dataframe or string. A table of variants containing variant allele frequencies per gene and per sample.}

\item{acn_save_path}{(optional) String. The output path (absolute path recommended) where to save the result.}

\item{return_sols}{(optional) Logical. Return the selected rascal solution.}

\item{return_S4}{(optional) Logical.}

\item{distance_decimal_places}{number of decimal places to compare to when finding
the minimum distance.}
}
\value{
A list containing: \cr
\enumerate{
\item A list of dataframes (one for each sample) OR a QDNAseq S4 object. \cr
\item Optionally, a dataframe of the rascal solutions. \cr
This function returns ACNs when a rascal solution is found, if one isn't, the sample is skipped.
}
}
\description{
CalculateACNs() calculates absolute copy numbers (ACNs) from the relative copy number profiles of one or more samples.
There are several included options by which to do this.
Note: If not providing a table of variant allele frequencies (VAFs) then 'mad' is the only method available. \cr

This function makes use of the rascal package in R and instructions can be found in the vignette: \cr
https://github.com/crukci-bioinformatics/rascal/blob/master/vignettes/rascal.Rmd  \cr
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{solutions <- "~/Documents/.../rascal_solutions.csv"
rcn.obj <- "~/repos/utanosmodellingdata/sample_copy_number_data/sample_filtered_cn_data.rds"
variants <- "~/Documents/.../allvariants.clinvar.cosmic.exons.csv"
save_path <- "~/Documents/.../rascal_ACN_segments.rds"
variants <- data.table::fread(file = variants, sep = ',')
variants <- variants \%>\% dplyr::rename(chromosome = chr,
                                       gene_name = genecode_gene_name)
outputs <- CalculateACNs(rcn.obj,
                         acn.method = 'maxvaf',
                         rascal_sols = solutions,
                         variants = variants,
                         return_sols = TRUE,
                         return_S4 = TRUE)
output <- CalculateACNs(rcn.obj,
                        rascal_sols = solutions,
                        variants = variants,
                        acnmethod = c('TP53', 'KRAS', 'BRCA1',
                                      'BRCA2', 'PIK3CA', 'PTEN'),
                        acn_save_path = save_path)
output <- CalculateACNs(rcn.obj,
                        rascal_sols = solutions,
                        acnmethod = 'mad')

}\if{html}{\out{</div>}}
}
