% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureplotting.R
\name{GenomicImportancePlot}
\alias{GenomicImportancePlot}
\title{Make a Genomic Importance plot}
\usage{
GenomicImportancePlot(importance, input_df, ref_genome = "hg19")
}
\arguments{
\item{importance}{\emph{list.} A list of vectors. \cr
There should be two vectors for each signature, one containing p-values and the other coefficients from a regression model.}

\item{input_df}{\emph{data frame.} A data frame with rownames to be used as column names. Ex. the input for \link{FitGIModelsCNtoS}.}

\item{ref_genome}{\emph{character.} The reference genome used in creating this dataset. ex. "hg19"}
}
\value{
A ggplot object.
}
\description{
This function processes p-values and coefficients from a list, transforms them into a long format,
and produces a plot highlighting regions of the genome accordingly.
}
