% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFilteredCNSummaryTable.R
\name{MakeSummaryTable}
\alias{MakeSummaryTable}
\title{Makes summary tables}
\usage{
MakeSummaryTable(
  CNobj,
  snames,
  lowT,
  highT,
  pL,
  pG,
  prop,
  find_peaks = FALSE,
  save_path = FALSE
)
}
\arguments{
\item{CNobj}{An object. A a ballgown object containing Copy Number data.}

\item{snames}{A vector. A vector of sample IDs that are of interest.}

\item{lowT}{A whole number. A threshold below which there is copy number loss.}

\item{highT}{A whole number. A threshold above which there is copy number gain.}

\item{pL}{A float. A probability threshold (between 0 and 1), to which loss frequency is compared.}

\item{pG}{A float. A probability threshold (between 0 and 1), to which gain frequency is compared.}

\item{prop}{A float. A proportion threshold (between 0 and 1), to which loss and gain occurences are compared (separately).}

\item{save_path}{A string. A path (directory) to where segment tables should be saved. ex. '~/Documents/test_project'.}
}
\value{
Nothing.
}
\description{
MakeSummaryTable takes in a Copy Number object and writes a TSV table containing all the summaries.
It is the function that does the work.
}
