% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureplotting.R
\name{MixtureModelPlots}
\alias{MixtureModelPlots}
\title{Plot Mixture Model Components by Signature}
\usage{
MixtureModelPlots(signatures, components, sig_of_interest = 1, threshold = 0.3)
}
\arguments{
\item{signatures}{A dataframe with components along the y-axis and signatures along the x.}

\item{components}{A list of S4 objects belonging to the class 'flexmix'.}

\item{sig_of_interest}{(optional) A single integer corresponding the signature for which to make plots. \cr
In the absence of any value passed in for this parameter, plots for just the first signature will be returned.}
}
\value{
A list of ggplots.
}
\description{
This function (MixtureModelPlots), for an indicated CN-Signature (S), returns a list of six ggplots.
Each plot visualizes the mixture models used for each feature.
All mixture components are plotted, but those elevated for S are shaded.
}
