% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{SignatureExposuresPlot}
\alias{SignatureExposuresPlot}
\title{Create Heatmap of Signature Exposures}
\usage{
SignatureExposuresPlot(
  signatures,
  order = FALSE,
  transpose = FALSE,
  save_path = FALSE,
  obj_name = "sig_exposures_obj"
)
}
\arguments{
\item{signatures}{Dataframe. Expects a dataframe of signature exposures (rows) by samples (columns).}

\item{order}{(optional) Character vector. Defines the order in which samples will be plotted. \cr
This is particularly useful when plotting more than 1 heatmap next to one another. \cr
Allows samples to line-up horizontally. \cr
Example: \cr
\verb{c("CC-CHM-1341", "CC-CHM-1347", "CC-CHM-1355", CC-CHM-1361", "CC-HAM-0369", "CC-HAM-0374", "CC-HAM-0379", "CC-HAM-0383", "CC-HAM-0385")}}

\item{transpose}{(optional) Logical. If set to TRUE the function returns the order in which samples were plotted.}

\item{save_path}{(optional) String. Expects a path to a directory where the plot should be saved. (png)}

\item{obj_name}{(optional) String. Adds a tag to the end of the filename if saving the image. \cr
Only used if the save_path parameter is also set.}
}
\value{
A list of a ggplot2 object and vector of the ordered sample names.
}
\description{
Converts signature-per-sample data to a heatmap, optionally saves it as a png, and returns the ggplot.
Samples are sorted for display based on their maximum signature exposure.
Heatmap is plotted in the viridis colour-scheme.
}
