% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{SummaryCNPlot}
\alias{SummaryCNPlot}
\title{Generate Summary Copy-Number Aberrations Plot}
\usage{
SummaryCNPlot(
  x,
  main = "Relative Copy-Number Summary Plot",
  maskprob = 0.2,
  maskaberr = 0.1,
  gaincol = "blue",
  losscol = "red",
  misscol = NA,
  build = "GRCh37",
  ...
)
}
\description{
This function was copied and modified from another package (CGHbase).
The original code can be found here on github: https://github.com/tgac-vumc/CGHbase
The bioconductor page: https://bioconductor.org/packages/release/bioc/html/CGHbase.html
}
